backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
qrcode[pil]
python-dotenv